adapt and move fingerprint to tracker parent class
